EDA using Python for organizing information.
In this repository, you will find an exploratory data analysis of the Forbes 2000 company's data using the powerful data manipulation library pandas, the data visualization library Matplotlib, and Seaborn. The analysis includes a detailed examination of the data set to uncover patterns and trends and provides visualizations to help understand the data. The groupby function is used to group the data by various attributes, such as industry and country, to gain further insights. This repository is a useful resource for anyone interested in understanding the characteristics of the world's largest companies and learning more about data analysis using Pandas, Matplotlib, and Seaborn.