Add files via upload
This project is a data analysis of 40 countries over 14 years, focusing on 7 variables. The data has been collected separately for each country and pre-processed, including cleaning of missing values. Graphs have been plotted to visualize the data and the final dataset has been merged for further analysis in a research project. The code for data cleaning, pre-processing, and visualization is available on GitHub. The final formation of the data is also available in this repository.