Time Series ARIMA with Python
This project is a time series analysis of 3M & Co. stock prices using data from 2016 and 2017. The ARIMA model was implemented to predict the daily high price of the stock. The mean squared error of the model was 0.12 and it was able to explain 91% of the variation in the data. The project includes code for data cleaning, pre-processing, model implementation, and evaluation. The dataset used in this project is available on GitHub for replication and reference.